using NotFoundException in DeleteLeaveRequestCommandHandler

+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
@@ -18,7 +19,11 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
 
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
+
         var leaveRequest = await _leaveRequestRepository.GetById(request.Id);
+
+        if (leaveRequest == null)
+            throw new NotFoundException(nameof(leaveRequest), request.Idid);
         await _leaveRequestRepository.Delete(leaveRequest.Id);
         return Unit.Value;
